Made proper adjustments from the US 41 merge. Fixed bugs and compile errors. Everything works as it should now.
