+++ app/src/main/java/com/capstoneproject/arrivalnotification/TransporterActivity.java
@@ -8,6 +8,7 @@ import android.content.pm.PackageManager;
 import android.hardware.camera2.CameraAccessException;
 import android.hardware.camera2.CameraCharacteristics;
 import android.hardware.camera2.CameraManager;
+import android.location.Location;
 import android.os.Bundle;
 import android.support.v4.app.ActivityCompat;
 import android.support.v7.app.AppCompatActivity;
@@ -16,8 +17,13 @@ import android.widget.Button;
 import android.widget.TextView;
 import android.widget.Toast;
 
+import com.google.android.gms.location.FusedLocationProviderClient;
+import com.google.android.gms.location.LocationServices;
+import com.google.android.gms.tasks.OnSuccessListener;
 import com.google.zxing.integration.android.IntentIntegrator;
 import com.google.zxing.integration.android.IntentResult;
+
+import static android.Manifest.permission.ACCESS_FINE_LOCATION;
 
 public class TransporterActivity extends AppCompatActivity {
         public static final String EXTRA_MESSAGE = "com.example.myfirstapp.MESSAGE";
@@ -30,6 +36,8 @@ public class TransporterActivity extends AppCompatActivity {
         private final Activity actitvity = this;
         private TextView bar_scanner;
         private Button btn_camera;
+        private FusedLocationProviderClient lastKnownLocation;
+        private String latitude, longitude;
 
     @Override
     protected void onCreate(Bundle savedInstanceState)
@@ -38,13 +46,29 @@ public class TransporterActivity extends AppCompatActivity {
         //Set layout to transporter
         setContentView(R.layout.activity_transporter);
 
+        requestPermission();
+
         bar_scanner = this.findViewById(R.id.scanning_view);
         btn_camera = this.findViewById(R.id.btn_camera);
+        lastKnownLocation = LocationServices.getFusedLocationProviderClient(this);
 
         btn_camera.setOnClickListener(new View.OnClickListener(){
             @Override
             public void onClick(View v)
             {
+                if (ActivityCompat.checkSelfPermission(TransporterActivity.this, ACCESS_FINE_LOCATION) != PackageManager.PERMISSION_GRANTED) {
+
+                    return;
+                }
+                lastKnownLocation.getLastLocation().addOnSuccessListener(TransporterActivity.this, new OnSuccessListener<Location>() {
+                    @Override
+                    public void onSuccess(Location location) {
+                        if (location != null) {
+                            latitude = "Latitude: " + location.getLatitude();
+                            longitude = "Longitude: " + location.getLongitude();
+                        }
+                    }
+                });
                 cameraManager = (CameraManager) getSystemService(Context.CAMERA_SERVICE);
                 context = TransporterActivity.this.getApplicationContext();
                 openCamera();
@@ -93,7 +117,7 @@ public class TransporterActivity extends AppCompatActivity {
             if (res.getContents() == null) {
                 Toast.makeText(this, "Cancelled", Toast.LENGTH_LONG);
             } else {
-                updateText(res.getContents());
+                updateText(res.getContents() + "\n" + latitude + "\n" + longitude);
             }
         } else {
             super.onActivityResult(reqCode, resCode, data);
@@ -102,5 +126,9 @@ public class TransporterActivity extends AppCompatActivity {
 
     private void updateText(String getCode){
         bar_scanner.setText(getCode);
+    }
+
+    private void requestPermission() {
+        ActivityCompat.requestPermissions(this, new String[]{ACCESS_FINE_LOCATION}, 1);
     }
 }
